商品削除機能（Rubocop）

+++ app/controllers/items_controller.rb
@@ -36,10 +36,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    if item.user == current_user
-      item.destroy
-    end
-      redirect_to root_path
+    item.destroy if item.user == current_user
+    redirect_to root_path
   end
 
   private
